Removed the handler/site after reading the API doc about it.
I don't want to take on many things I don't understand at once.
It was replaced with #'compojure-app. So now app is processing
just the handler returned by compojure-app with no extra un-seen
middleware.

Also, core.clj changes are seen without restarting now because we're
passing the var as the argument to thread-first not the function itself.
I don't yet fully understand this. But it is talked about here:
http://www.learningclojure.com/2013/01/getting-started-with-ring.html